extract account_hash mod.
